requirements.txt
astroid==2.3.0
autopep8==1.4.4
cffi==1.14.0
cryptography==2.9
Django==2.2.5
django-cors-headers==3.2.1
django-guardian==2.1.0
django-restql==0.5.3
djangorestframework==3.10.3
Faker==2.0.4
gunicorn==20.0.0
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
pep8==1.7.1
Pint==0.11
psycopg2==2.8.3
pycodestyle==2.5.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.4.1
pylint-django==2.0.11
pylint-plugin-utils==0.6
pyPEG2==2.15.2
python-dateutil==2.8.1
python-decouple==3.1
python-json-logger==0.1.11
python-logstash==0.4.6
python3-logstash==0.4.80
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.3
typed-ast==1.4.0
wrapt==1.11.2
Pillow==6.1.0
django-cleanup==5.1.0
django-extra-fields==3.0.2
PyPDF2==1.26.0
fiscalyear==0.3.0
django-filter==2.2.0